+++ controls/admin.py
@@ -1,7 +1,7 @@
 import csv
 from django.contrib import admin
 from django.http import HttpResponse
-from .models import Statement, Element, ElementControl, System, CommonControlProvider, CommonControl, ElementCommonControl
+from .models import Statement, Element, ElementControl, System, CommonControlProvider, CommonControl, ElementCommonControl, Poam
 from guardian.admin import GuardedModelAdmin
 
 class ExportCsvMixin:
@@ -24,12 +24,12 @@ class ExportCsvMixin:
     export_as_csv.short_description = "Export Selected as CSV"
 
 class StatementAdmin(admin.ModelAdmin, ExportCsvMixin):
-    list_display = ('id', 'sid', 'sid_class', 'producer_element')
+    list_display = ('id', 'sid', 'sid_class', 'producer_element', 'uuid')
     actions = ["export_as_csv"]
-    readonly_fields = ('created', 'updated')
+    readonly_fields = ('created', 'updated', 'uuid')
 
 class ElementAdmin(GuardedModelAdmin, ExportCsvMixin):
-    list_display = ('name', 'full_name', 'id')
+    list_display = ('name', 'full_name', 'id', 'uuid')
     actions = ["export_as_csv"]
 
 class ElementControlAdmin(admin.ModelAdmin, ExportCsvMixin):
@@ -53,6 +53,16 @@ class ElementCommonControlAdmin(admin.ModelAdmin, ExportCsvMixin):
     list_display = ('element', 'common_control', 'oscal_ctl_id', 'id')
     actions = ["export_as_csv"]
 
+class PoamAdmin(admin.ModelAdmin, ExportCsvMixin):
+    list_display = ('id', 'statement')
+    actions = ["export_as_csv"]
+
+    def uuid(self, obj):
+        return obj.statement.uuid
+
+    def consumer_element(self, obj):
+        return obj.statement.consumer_element
+
 
 admin.site.register(Statement, StatementAdmin)
 admin.site.register(Element, ElementAdmin)
@@ -61,3 +71,4 @@ admin.site.register(System, SystemAdmin)
 admin.site.register(CommonControlProvider, CommonControlProviderAdmin)
 admin.site.register(CommonControl, CommonControlAdmin)
 admin.site.register(ElementCommonControl, ElementCommonControlAdmin)
+admin.site.register(Poam, PoamAdmin)