+++ siteapp/models.py
@@ -1152,3 +1152,14 @@ class Invitation(models.Model):
 
     def get_redirect_url(self):
         return self.target.get_invitation_redirect_url(self)
+
+class Support(models.Model):
+  """Model for support information for support page for install of GovReady"""
+
+  email = models.EmailField(max_length=254, unique=False, blank=True, null=True, help_text="Support email address")
+  phone = models.CharField(max_length=24, unique=False, blank=True, null=True, help_text="Support phone number")
+  text = models.TextField(unique=False, blank=True, null=True, help_text="Support desription at top of page")
+  url = models.URLField(max_length=200, unique=False, blank=True, null=True, help_text="Support url")
+
+  def __str__(self):
+    return "Support information"